Refactor import paths for music page and menu bar components
